+++ pkg/modes/acp/acp_test.go
@@ -1271,6 +1271,7 @@ func (n *noopBridgeAPI) ReportProgress(_ string)         {}
 func (n *noopBridgeAPI) Introspect() session.Introspection {
 	return session.Introspection{}
 }
+func (n *noopBridgeAPI) RestartSession(_ string) error { return nil }
 
 // writeCommandExtScript writes a Python extension script that:
 //   - responds to the init handshake with a "greet" command